requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
pytz
phonenumberslite
twilio
python-dotenv>=1.0.0
django-cors-headers>=4.3.1
